chore: Refactor admin pages to have a better separation of routes
