+++ packages/earth-admin/src/components/places/index.ts
@@ -17,6 +17,7 @@
   specific language governing permissions and limitations under the License.
 */
 
-import PlaceDetails from './PlaceDetails';
-
-export { PlaceDetails };
+export * from './place-edit';
+export * from './place-intersections';
+export * from './place-metrics';
+export * from './place-sidebar';
