+++ packages/earth-admin/src/components/places/place-intersections/index.ts
